final correction
after further research i found that the plugin I added at the pom.xml was only for mvn site. I don't need it for now. So I suppressed it.

The goal was to obtain 60% test coverage on jacoco report.

Just needed to suppress executable and run mvn clean, mvn verify and mvn test.